Add option to reduce LQR control input size
A system may have an input size of 'm', but a realizable control input
size less than 'm'. This can be set with the LQR input weight matrix
'R'. By setting a diagonal entry of 'R' to zero, the corresponding entry
in 'u', the control input vector, will always set to zero. This process
reduces the size of the system input in all LQR computations.

For example, with 'm' = 4 and
    R = [ *, 0, 0, 0 ]
        [ 0, 0, 0, 0 ]
        [ 0, 0, 0, 0 ]
        [ 0, 0, 0, * ]

The LQR controller will always return a control input vector
    u' = [ *, 0, 0, * ]