Integrate Bicycle auxiliary states
Add auxiliary states 'x', 'y', 'theta' which are used only for
visualization, where:
    'x' is the x-position of the rear wheel contact point
    'y' is the y-position of the rear wheel contact point
    'theta' is the pitch angle of the rear frame

The trajectory of the auxiliary states can be computed using Bicycle
member function 'x_aux_next' to update the auxiliary states at each time
step. It is expected that the user call this function after 'x_next' as
neither the state nor auxiliary state are stored in the Bicycle object.
The auxiliary states are calculated using a boost explicit integrator
with the same time step use in the discrete time system formulation.