Replace std::pow(x, 2) with x*x
In commit 3325447bb881bc1707f1f0dda4f18a7bfe8ef0a1 std::pow(x, 2) was
changed to std::pow(x, static_cast<real_t>(2.0)) to prevent a double
from being returned when using C++11. Refer to
http://en.cppreference.com/w/cpp/numeric/math/pow.

As the second argument type is not an integer value, it is unlikely for
the compiler to optimize this expression as x*x so it is done manually.