Fix pure virtual call in Bicycle constructor
set_state_space() is called in the Bicycle constructor. As this is
declared a virtual function, the method of the derived class is invoked.
Since construction of the derived class starts after the Bicycle
constructor is executed, the call to set_state_space() will fail as it
attempt to use the derived class version.

Instead, Bicycle::set_state_space() is called in the constructor and the
derived class constructor is responsible for other initialization
necessary.